Add support for sticky sessions
The MQTT traffic is load balanced using nginx and
a vernemq cluster. Since mqtt state is a lot to maintain
we need to provide support for sticky sessions. Sticky sessions
can be maintained using a custom load balancing algorithm which
uses ClientID to direct requests. It is based on nginScript
support from nginx which allows writing a js script to direct to
the right server node.